requirements.txt
pandas
numpy
torch
sklearn
matplotlib
click
pyyaml